Clean: Relocate STEX GridAccessor
