+++ src/foundation/CMakeLists.txt
@@ -28,6 +28,7 @@ SET(
         MatT.h
         PriorityQueue.h
         Grid.h
+        GridAccessor.h
         ValueConv.h
 )
 SOURCE_GROUP("Sources" FILES ${sources})
